% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpartforest.R
\name{rpartforest}
\alias{rpartforest}
\title{Build a random forest with base learners grown from rpart}
\usage{
rpartforest(
  y,
  x,
  test,
  mtry,
  ntrees,
  getlm = F,
  contrl = rpart.control(minsplit = 5, minbucket = 1, cp = 0.001)
)
}
\arguments{
\item{y}{\itemize{
\item outcome variable
}}

\item{x}{\itemize{
\item explanatory covariates
}}

\item{test}{\itemize{
\item testing data
}}

\item{mtry}{\itemize{
\item mtry tuning parameter
}}

\item{ntrees}{\itemize{
\item number of trees within the forest
}}

\item{getlm}{\itemize{
\item boolean to indicate whether linear model representation should be computed
}}

\item{contrl-}{control parameters for rpart; refer to that documentation for details}
}
\value{
A matrix of predicted values from each base learner as well as
training and testing ANOVA matrices
}
\description{
This is a rudimentary implementation of Random Forest using rpart.
It primarily serves as a proof-of-concept to utilize the "treelm" function
to enable the representation of a Random Forest as a collection of design matrices
}
\author{
Ian Frankenburg
}
