% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treelm.R
\name{treelm}
\alias{treelm}
\title{Tree to linear model}
\usage{
treelm(model)
}
\arguments{
\item{model}{a fitted rpart object or a fitted ctree model}
}
\value{
A list containing the tree topology and design matrix
\itemize{
\item X - the ANOVA design matrix for the linear model
}
}
\description{
The treelm function transforms the fitted rpart or ctree object into a
design matrix representing the tree structure. The default encoding follows from
Michael LeBlanc & Robert Tibshirani (1998) Monotone Shrinkage of Trees,
Journal of Computational and Graphical Statistics,
7:4, 417-433, DOI: 10.1080/10618600.1998.10474786
}
\details{
Represent a fitted regression or classification tree as an ANOVA model
}
\author{
Ian Frankenburg
}
